Fill standard questions with profile values
Allow two new values:

OKTA_USERNAME for the default username of a profile
OKTA_AUTH_METHOD (none, question, sms, google, okta-verify are allowed values) to select the authentication method.

If these values are set on a profile, do not prompt in interactive mode but use these values directly.

If only a single role is presented for selection in interactive mode, select the role automatically.